+++ back-end/src/main/java/com/tdd/backend/car/CarRepository.java
@@ -1,4 +1,4 @@
-package com.tdd.backend.option.render;
+package com.tdd.backend.car;
 
 import java.util.Optional;
