+++ back-end/src/main/java/com/tdd/backend/option/render/RenderService.java
@@ -5,6 +5,7 @@ import java.util.List;
 
 import org.springframework.stereotype.Service;
 
+import com.tdd.backend.car.CarRepository;
 import com.tdd.backend.option.Category;
 import com.tdd.backend.option.data.OptionDto;
 import com.tdd.backend.option.data.OptionResponse;
